[BUG] QTI in: TextEntry - don't use `expectedLength` to set max chars
This change applies to `TextEntryInteraction`, which maps to Learnosity's
`clozetext` question type.

We currently use `expectedLength` to set the max character length on the
response containers. This has the side-effect of defining the width of the
containers to fit the max number of characters specified (a behavior of
Questions API.) This can in fact be undesirable, as the IMS spec does not
specify that it is intended to be used this way (https://www.imsglobal.org/question/qti_v2p0/imsqti_infov2p0.html#element10300).
As a result, QTI content may mistakenly pass a smaller `expectedLength`
than the possible valid answers require. On the other hand, it is still
desirable to use `expectedLength` to help define the response width.

This change modifies the `TextEntryInteraction` type to _not_ use
`expectedLength` for the character max length by default, and adds
a flag to turn it on. Additionally, it adds support for setting the
`response_container.width` attribute based on the `expectedLength`,
and makes the default width `null` to allow Questions API to determine
the width itself.