Fix empty responseDeclaration prevents responseProcessing execution
This commit revert the ability to bypass responseProcessing execution
based on an empty responseDeclaration block. According to the QTI docs,
the responseDeclaration block can be empty.

The original check got introduced in commit
3a69bfa65a7dab294c7b98f49baa3d0729ea75b6 as an attempt to handle cases
where there is no correct responses.

This change therefore may break an old edge case, but the fix should be
in a more isolated section instead of blanket skipping the processing of
responseProcessing rules.

There is also a minor refactor to make sure we derive the correct
expression types before continue processing.